+++ docs/references/changelog.md
@@ -21,6 +21,8 @@ versioning][semver].
 
 - `kele-context` and `kele-resource` now wait on kubeconfig sync completion to
   finish if one is currently in progress
+- Fixed a bug where force-enabling or force-disabling `kele-mode` (via either
+  `(kele-mode 1)` or `(kele-mode -1)`) resulted in errors
 
 ## 0.3.0